last push

+++ FSDProject/Account.cs
@@ -8,7 +8,7 @@
         public string DateOfBirth { get; set; }
         public string Phone { get; set; }
         public string Address { get; set; }
-        public int Funds { get; set; }
+        public double Funds { get; set; }
         public int IsAdmin { get; set; }
     }
 }
